+++ LocalAgent_v0.2.12/Properties/AssemblyInfo.cs
@@ -1,0 +1,10 @@
+﻿using System.Reflection;
+using System.Runtime.InteropServices;
+[assembly: AssemblyTitle("VisionQC Local Agent")]
+[assembly: AssemblyDescription("Local VPDL Runtime bridge for VisionQC Web")]
+[assembly: AssemblyCompany("VisionQC")]
+[assembly: AssemblyProduct("VisionQC Local Agent")]
+[assembly: AssemblyVersion("0.2.12.0")]
+[assembly: AssemblyFileVersion("0.2.12.0")]
+[assembly: ComVisible(false)]
+[assembly: Guid("8e8d7e74-77cb-4e14-8b72-57009b19ed40")]